[Registry-Frontend] Serialization and deserialization, save/load removed structs
Serialization and deserialization: Now the registry frontend have functions to serialize the configuration to a vector of uint8_t for the backend to save. Also save method that will save the configuration to the backend. Added load method that de-serializes the saved vector (for now the same since no get from the backend), performs checks and loads into the map the configuration.
Removed structs: The specific structs are removed since they will be used by the OBSW and not in the Boardcore sw.
Tests: Added arm/disarm tests, added a serialization/deserialization save/load test.
Reserve: Added reserves to the dynamic data structures to limit the allocations.
Minor fixes: coordinates was erroneously in uint32 and not float.

+++ src/shared/utils/Registry/TypeStructures.h
@@ -78,11 +78,10 @@ struct UInt32Type : RootTypeStructure<uint32_t, T_ENUM>
  */
 struct Coordinates
 {
-    uint32_t latitude;
-    uint32_t longitude;
+    float latitude;
+    float longitude;
 
-    explicit Coordinates(const uint32_t setLatitude,
-                         const uint32_t setLongitude)
+    explicit Coordinates(const float setLatitude, const float setLongitude)
         : latitude(setLatitude), longitude(setLongitude)
     {
     }
